Deleted horrible exception

+++ web/helper.py
@@ -165,10 +165,8 @@ def parse(room, users, impath):
     
     del draw
 
-    try:
+    if os.path.exists(os.path.join(basedir, 'images', str(room.id) + '_map.jpg')):
         os.remove(os.path.join(basedir, 'images', str(room.id) + '_map.jpg'))
-    except FileNotFoundError:
-        pass
     room_map.save(os.path.join(basedir, 'images', str(room.id) + '_map.jpg'))
